requirements.txt
streamlit
firebase-admin
langchain
langchain-openai
openai
tavily-python>=0.5.4
pypdf
langchain-google-community